properly encapsulate overlay hack and properly dispose overlays
Add the function `magit-put-face-property' as a substitute for
`magit-put-text-property' to work around a limitation mentioned
in info node `(elisp)Overlay Properties':

  Currently, all overlays take priority over text properties.
  Please avoid using negative priority values, as we have not
  yet decided just what they should mean.

We already did the same before, but re-implemented that in every case
in which it was needed.  Unfortunately each of these implementations
contained the same severe bug: the overlays were not removed when the
diffs got updated.  Overlays negatively affect performance and so these
buffers got slower over time.  Now these overlays "evaporate" when they
become empty, i.e.  they are automatically removed.