use overlays to propertize various parts of logs
Ever wondered why faces used for refs such as branch names have a
border?  It is because of this limitation mentioned in info node
`(elisp)Overlay Properties':

  Currently, all overlays take priority over text properties.
  Please avoid using negative priority values, as we have not
  yet decided just what they should mean.

So when a section is highlighted by changing the background (the
default), then all background colors things like reflabels that are
located inside the section loose their background color.  It seems
someone decided that just the foreground color being different from
surrounding text is not enough and also added a border to the faces
used for such things.  This has been reverted in some color themes
in others it has not.  After this change it is solely a matter of
taste, not visibility whether someone wants to keep the borders.

The borders will probably be removed from the defaults of these
faces eventually.  But this isn't urgent.