Set this-command in magit-key-mode-command
Some library (ido) could have problem when this-command is set to
something that is not a symbol, and when a key is pressed, its binding
is put into this-command. In magit-key-mode, lambda expression are bind
to key, and this could confuse ido.

As a side effect, last-command will be set to the real command that have
be run, and not to lambda expression.

The problem with ido is probably a bug in Emacs.