move definitions of functions and variables used for diff washing
In preparation of major refactoring move definitions of functions and
variables used for diff washing.  The old location has never made
much sense.

Only move the functions and variables that implement "diff" and "hunk"
washing, and leave the "raw diff" functions where they are now.  They
are not going to be refactored.  This extra washing cycle will be
removed without a replacement because it has no benefits but many
downsides.