magit-git-executable: use executable-find when initializing value
This hopefully give a slight performance bust.

The downside is that the headers in the process buffer get longer
and that the "interesting parts" might end up being truncated.

Still I hesitate to add an option to hide the full path and the
`magit-git-standard-options'; when instructing users to provide
the information required to debug some issue there would be one
more thing that we would have to make sure they actually did.

If `executable-find' fails to find Git then fallback to "git".