magit-git-executable: use git.exe on Windows
On Windows, explicitly use the "git.exe" executable instead of "git".
If that fails accept any "git".

It has been reported that this gives a considerable speedup.  I don't
know Windows but it appears that when asking for a "foo" one might get
a "foo.exe" or a "foo.cmd" or a "foo".  It also appears that "git.cmd"
is earlier on the `exec-path' than "git.exe", or that the former is
preferred for some other reason.