magit-run-{git-gui{,-blame},gitk}: use call-process
Start the subprocess using `call-process' with argument DESTINATION
being 0.  I.e. start the subprocess and immediately return 0 without
caring what happens to the subprocess afterwards.  Previously we used
`start-file-process' but that had the disadvantage that the process
object still existed when the subprocess was still running when Emacs
was being quit, needlessly forcing the termination of the subprocess.