fixup diffstat sections in more civilized manner
Git may truncate filenames in diffstats, and because we use these string
for more than just showing them to the user we have to fix that.

The real filenames become available once when the diffs are washed, at
which point store these filenames in the diffstats section objects too.
For that to work we have to somehow pass the diffstat sections to the
code that washes the diffs.  Previously that was done using the variable
`magit-diffstat-cached-sections'.

By merging `magit-wash-diffstat' into `magit-wash-diffstat' we can now
do the civilized thing and pass the information using function arguments.