magit-wash-hunk: turn the first line into a proper heading
Previously the first line of a hunk looked like it was a heading but as
far as the section code was concerned it was actually just dressing up
as one.  The fact that it still worked like a section heading does not
contradict this statement, it just means that there is code within the
section handling that can deal with such broken sections, these kludges
could now be removed.  (We first have to make sure that no other section
remain that are broken in the same way, though.)

This commit turns the first line into a proper headings but due to the
inflexibility of `magit-with-section' with regards to how headings are
painted we have to delay the creation of the overlay until after the
`magit-with-section' has returned.  (The overlay is also a hack which
in turn is necessary because Emacs doesn't support negative overlays.)