dispose of unnecessary diff washing cycle
Previously staged and unstaged changes as displayed in the status buffer
were washed in two washing cycles.  All other diffs are and always were
washed in just one washing cycle, like everything else is too.  This was
either a historic accident or a failed attempt at improving performance.

If it was the latter then the motivation probably was that this allowed
delaying the insertion and washing hidden diffs until the got expanded.
However this actually decreases performance because it means Git has to
be called several times (at least on Windows where starting a process
seems to come with a huge overhead).

Getting the diff from Git once isn't what takes a long time, its Magit's
inefficient parsing and painting of the diff.  These steps have already
been made faster and we will be further improved.  If it turns out that
we cannot speed up the diff washing enough even for huge diffs then we
can still delay *washing only*.  To do that we don't have to also delay
the insertion of the diff.