magit-wash-diff: delete Git's diff headers
Replace the file and hunk diff headers after they have been parsed and
the Magit equivalents have been inserted.  This change is motivated by
to things: the standard headers are redundant and therefor only make the
diffs noisier, and it improves performance to remove these headers and
therefor also not having to paint them using overlays.

There is a downside too: it is no longer possible to save a Magit diff
buffer to a file and then pass that patch to Git or something else that
understands unified diffs.