magit-rewrite-{set-commit-property,finish-step}: fix regressions
Make `magit-rewrite-set-commit-property' always return t so that the
`magit-section-action' hooks `magit-{apply,cherry-pick,revert}-hook'
run just the rewrite case, and not the default `commit' case too.

In `magit-rewrite-finish-step', do not use `magit-cherry-pick-commit'.
That function does a refresh before we had a chance to call
`magit-rewrite-set-commit-property'.  If cherry-picking a commit fails,
then say so.