magit-{insert,put-face-property}: don't specify overlay priority
We cannot specify the default priority because the overlays created to
refine hunks (by code in `smerge-mode.el') also are created without the
priority being explicitly specified.  When the priority is undefined,
then it defaults to 1.  In our case there are several overlays with that
priority at the same buffer position, but it seems the correct on wins.

Apparently if there are multiple overlays with the same priority then
the one which was created later is prioritized.  It so happens that this
is what we need here, but we just got lucky and this should be fixed in
Emacs.