magit-wash-hunk: refine hunk outside of magit-with-section
I repeat myself but "due to limitation in the section code", we have to
delay refining the hunk until after the call to `magit-with-section'.
More precisely the "@@..." header has not been reinserted yet when the
BODY of this macro is evaluated, and so any textual search for the
beginning of the hunk causes an error.