magit-define-section-jumper: absorb magit-goto-section-at-path
Merge `magit-goto-section-at-path' into its only caller
`magit-define-section-jumper'.  This fixes a bug as the former did
not return nil when the section could not be found, but the latter
expected it to do so.  Instead of just fixing that, merge because
the name `magit-goto-section-at-path' is to generic for a function
which only knows about sections that are direct children of the
root section.