magit-insert-status-tags-line: use only colors, not verbs
Avoid uncertainty about whether the tag or the branch is the subject
when displaying the current and next tags in the header of the status
buffer, by using just colors to indicate which side has the additional
commits. Previously we used the verbs "ahead" and "behind", and which
one had to be used depended on whether the branch or the tag was
considered the subject of the "sentence".