sanitize some faces
Remove the face `magit-header' instead make `magit-section-title'
inherit from `header-line' directly, without the redirection through
`magit-header'.

The removed face was described as "Many Magit faces inherit from this
one by default", but for those faces that are *not* used by "headers"
aka "titles", doing so was wrong.  So do *not* change those faces to
inherit from `magit-section-title' and instead use the same colors as
for the respective `magit-log-head-label-*' faces, but don't put a
`:box' around them.

There is more wrong with these and other faces, but that will be
fixed later.