clarify distinction between a section's type and value
Originally there were these three related but distinct slots: `type',
`info', and `header'.  The last one was unnecessary and only existed
because someone got confused about the distinction between the other
two and was removed when I last refactored the section code.

Now the time has come to make the distinction between the "type" and the
"info" more clear.  We start by using the term "value" instead of
"info", because that's what it is.

There are two categories of sections: those that represent a thing, and
those that serve as containers for other sections.  A section can fall
into both categories.

Every section has a type, a symbol.  Sections that represent a "thing",
such as a file or hunk also have a value, usually a string (but that is
not enforced).  Sections that only serve as containers do not have a
value (i.e. the value is nil).

Together the type and value of a section and the types of recursively
all parent sections can be used to uniquely identify a section within
a buffer.  For container-only sections the types are sufficient because
there always is only one section of a particular container-only
type.  (That is a convention and isn't enforced.)  Meanwhile there can
be many e.g. `commit' sections and to tell them apart the values (here
the commit hashes) also have to be compared.

Sadly someone got confused along the way and started to use the just
the values (then more ambitiously called "info") to identify sections.
Since many sections don't actually have a value the result was that
for sections that would otherwise not have a value, the type was stored
in the `value' slot as well.

This commit adds function `magit-section-ident' which returns a list of
conses that can be used to uniquely identify a section as described.
Function `magit-find-section' is changed to use that value, instead of
using the value-except-when-it-actually-is-the-type.  For sections that
have no value the value of the `value' slot now is nil.