rename section struct slot BEGINNING to START
This causes the accessor `magit-section-beginning' to be renamed to
`magit-section-start'.  I'll admit it, I am doing this just because
"start" is shorter than "beginning".  Because the section code is
currently being refactored heavily anyway, doing so now causes only
little additional harm (in form of third-party code that might have
to be adapted), if any.