rename section struct slot CONTENT-BEGINNING to CONTENT
Currently this slot always holds a marker (though sometimes it points
somewhere strange and cannot be relied on).  Starting in subsequent
commits this slot will be used to hold either such a marker, nil (to
indicate that the section has no body), or a function (which is then
used to actually create the body on expansion).