magit-with-section: insert the heading before the body
Remove the HEADING argument from `magit-with-section' and instead expect
the callers to insert the HEADING themselves, using the new
`magit-insert-heading'.  Besides inserting the heading text in a fashion
similar to how the more general `magit-insert' works, this function also
sets the sections CONTENT slot to the position after the text previously
inserted.  The default value of that slot now is nil, indicating that
the section isn't divided into a heading and a body, i.e. it cannot be
collapsed.

Previously `magit-with-section' did evaluate its BODY to insert the
sections body text, and only then went back to the beginning of the
section and inserted the heading it had received as a separate argument.
This was done because when HEADING ending with ":" indicates that the
number of child sections shall be shown in the heading.  We no longer
delay inserting the heading until we know how many children there are.
Instead we insert the heading without the count and then later go back
and insert the count.  Inserting the heading later than the body
complicated things a lot and lead to difficult to fix bugs.