magit-cancel-section: new function
Previously we depended on the return value of `magit-with-section'
being nil to trigger the cancellation of the section currently being
inserted.  Now we use the new function `magit-cancel-section' which
does a non-local exit from a catch setup in `magit-with-section'.

Doing it this way is not necessary at this point yet, but it does give
us additional flexibility.  I.e. code that inserts a section no longer
has to be lexically within `magit-with-section's BODY.  We will use that
in a subsequent commit.