rename magit-with-section to magit-insert-section
Because this is a macro that does "something with something" the term
"with" actually does make sense.  However the "something it does" is
different from that that macros that are part of Emacs do.  We do not
make something current that already exists and we also don't create
something which is disposed of again after the macros BODY.  I.e. our
macro is neither like `with-current-buffer' nor like `with-temp-buffer'.

The side-effect of inserting a section is the raison d'être of this
macro.  Reflect that in the name.