magit-git-wash: new function
Separate the task of inserting and washing output from a command from
the task of creating a section.  This in a sense was already the case
previously in that `magit-cmd-insert-section' made use of the macro
`magit-insert-section' to do the creation part.

Replace `magit-cmd-insert-section' and `magit-cmd-insert-section' with
`magit-git-wash'.  It is just a thin wrapper around `magit-git-insert'
which additionally calls some function to wash the inserted text.  It
also knows how to cancel the section currently being created if either
Git produces no output or washing the output completely.  That is
different from what `magit-cmd-insert-section' which never canceled
before calling the washer.

Code that previously used `magit-git-insert-section' now has to
explicitly use `magit-insert-section', `magit-insert-heading', and
`magit-git-wash'.  That usually means that it takes four lines of code
instead of tree to do so, but it also makes this more flexible and
prettier to look at.

There is no substitute for `magit-cmd-insert-section', i.e. we now only
provide a "insert-and-wash" function for Git but not arbitrary commands,
because, well that's what Magit is about.  There were only two direct
uses of `magit-cmd-insert-section', one in `magit-stgit.el' and one in
`magit-topgit.el'.  These uses are now replaced with a bit of code
duplication.