improve process sections
Create all process sections "expanded".  Previously we did initially
collapse sections unless the *magit-process* buffer was already visible
in the current frame.  That was inconvenient because one might switch
to that buffer to see why something is taking long, only to first have
to expand the respective section.

Never-the-less it is useful to hide sections whose output the user
"probably doesn't want to see" in most cases.  So when a process returns
with exit code then the section is automatically collapsed.

Likewise inserting the exit code into the process sections buffer is now
done in a more reasonable way, without ending up putting point somewhere
inside the heading, instead of at its beginning.

Also do not put an extra empty line between sections and remove some
other kludges.  Now that headings are inserted before their bodies we no
longer need to do that.