there are no items, just sections and things
Until now the term "item" was used as a second order observation.  It
denotes that we do *not* differentiate between the gui element and the
object it represents.  The former is not a useful distinction and should
be avoided by removing "item" from symbol names and documentation.

The above actually makes sense, but unless you have read your share of
Luhmann, the information is lost.  So let's try again, using more words.

The main gui elements in Magit buffers are called "sections".  Sections
always have a "type", and either represent some object, in which case
they also have a "value", and/or serve as containers for other sections.

In Magit the object represented by some gui element could be something
like a file, commit, or hunk.  In Emacs speak the term "thing" is
usually used for an object that generally could be categorized, when we
don't actually want to or cannot make that distinction.  We don't use
"item" for that.

When we name a command `magit-VERB-item' we are neither saying "VERB the
thing represented by this gui element", nor "VERB this gui element".  We
are saying "VERB (either the gui element or the object it represents)".
So `*-item' does not make it any more obvious whether something is done
with the gui element or the represented object.  What it emphasizes is
that we *don't* make that distinction, i.e. the distinction it makes is
that we don't make the distinction that could actually be useful.  That
is less useful than not making a distinction at all.

Maybe "item" used to be a synonym for exactly one of "section" or
"thing"; but if that ever was so, then that meaning has long been lost.
Contemporary uses include "highlight the item at point", but also "open
the file whose name can be seen at point".

So, lets rename symbols whose names used to contain "item", in most, but
not all, cases by simply removing the "-item" suffix.  Also remove
"item" from doc-strings, where it is replaced with either "section",
"thing", or in some cases the actual type or types of the "things", as
appropriate in the individual cases.