improve how buffers are saved on status refresh
Improve how buffers are autosaved when the status buffer is refreshed
and before a checkout.  We do that more often now, i.e. every time the
status buffer is refreshed so we should stop saving totally unrelated
files.  Remove option `magit-save-some-buffers-predicate' that gave
masochistic users the possibility to be constantly asked "are we there
yet".

Rename the option `magit-save-save-some-buffers' and the function by
the same name to `magit-save-repository-buffers'.  Add the new wrapper
`magit-maybe-save-repository-buffers'.  Remove the variable
`magit-save-some-buffers-topdir' we "cache" that information more
gracefully now.  Remove code and documentation that looks like it was
just copied from the generic `save-some-buffers' implementation but is
not useful here.