use the diff buffer to show stashes
Originally we used a separate buffer for stashes
`magit-stash-buffer-name', some time ago we switched to using
`magit-commit-buffer-name' which was a mistake, instead we should use
`magit-diff-buffer-name' because `magit-diff-stash' is a "diff" variant.

Change that in `magit-diff-stash' and in `magit-show-or-scroll' (which
was still using the stash specific buffer and thus never succeeded).
Remove `magit-stash-buffer-name' variable.