magit-{stgit,topgit}: don't use action hooks
Do not use "action" hooks such as `magit-visit-hook' run by the command
`magit-visit' to add "specialized variants" of these actions for section
types added by extensions.

Such hooks are implemented using `magit-section-action' and I think that
adding this extension mechanism was a mistake in the first place.  Emacs
already has an extension mechanism that can be used to do this sort of
thing: the `keymap' text property.

Replace all uses of action hooks in `magit-stgit' and `magit-topgit by
using that `keymap' property for the text that is part of such
additional sections.

Both extension mechanisms have advantages and disadvantages.  Most
importantly the use of the `keymap' text property means that it is no
longer possible to change just one keymap to use another key for an
action, not only for core commands but also those defined by extensions.
On the other hand there now is a place where users can look to get a
list of all "special commands for a certain section type", the
respective keymap.