magit-topgit: don't overload core commands
Previously `magit-topgit-mode' added specialized functions to four
different command-specific hooks run by fundamental commands that are
part of Magit's core.  These commands did only run these hooks so that
`magit-topgit' could override the behavior of these commands.  I think
the behavior should never be overwritten in such a way.  And the core
of Magit certainly shouldn't be obfuscated to accommodate an extension
hardly anybody uses (280 downloads from Melpa, Magit itself has 57000).

Regardless of whether doing what we used to do is okay or not, we don't
have to do so because there is a better option: use a popup just for
TopGit commands.  Being parasites the TopGit "commands" where not
actually commands but regular functions, so that has to be fixed too.