magit-push: absorb magit-push-dwim
For a while we had a command `magit-push-dwim' separate from
`magit-push' for two reasons:

`magit-push' used to run hook `magit-push-hook' allowing `magit-topgit'
to override its behavior.  Since the previous commits we no longer use
it that way.

Since that hook existed anyway I decided to rename `magit-push's legacy
implementation to `magit-push-dwim' and add a new `magit-push' that did
run that by default.  This was done in preparation of the still
outstanding deprecation of the legacy implementation, which I think is
fairly bad, because it overloads pushing with only marginally related
functionality and departs to much from what Git does.

However there likely are some users who disagree and I intended to keep
the old implementation which for a while was known as `magit-push-dwim'
under the name `magit-push-oldschool', it is now known as just
`magit-push' again and is still going to be deprecated eventually.