magit-topgit-wash-topic: keep flags
Refactor `magit-topgit-wash-topic', making in readable and shrinking
it considerably.  In the process stop losing information (by no longer
removing the flag characters, only two out of seven were and still are
represented with face attributes), remove the unnecessary and ugly
`magit-topgit-parse-flags' and add the face `magit-topgit-empty'
(instead of hardcoding the appearance).