magit-insert-status-headers: new function
Add function `magit-insert-status-headers', replacing
`magit-insert-status-local-line', magit-insert-status-remote-line', and
`magit-insert-status-head-line'.  Now there are only two lines, the
"local" and the "remote" line, and they both look more like the removed
"head" line looked.  Or in other words the tip of the checked out branch
and the branch being tracked by that are now displayed in a fashion
similar to what would be shown in a log.