Provide option to avoid tracking branch prompt
Introduce a new customize option,
`magit-offer-to-create-tracking-branches', which controls whether or not
the user will be prompted to create a local tracking branch when
checking out a remote branch. Some users may prefer to *never* create
local tracking branches in this way, so the additional step when
checking out a remote branch might be irritating.