magit-add-change-log-entry{,-other-window}: move to magit-extras.el
Users and projects that have fully embraced Git don't use ChangeLog
files, they use commit messages.  Since their are users and projects
that have not done so, the functions `magit-add-change-log-entry' and
`magit-add-change-log-entry-other-window' still do server a purpose.
But these functions are not essential to Magit and are therefor moved
to `magit-extras.el'.