Add status buffer inserter for oneline graph
Sometimes it's nice to get a full graph view of recent commits in the
repository you're currently working on. Current status buffer inserters
are available for showing recent commits that *you* are working on but
nothing to just show recent commits in the repository.

Add an inserter that shows `n' commits with

    git log --graph --oneline --decorate --color

Since this inserter is constantly asking for --graph output, optimize
the computation of recent commits. For large repositories it's much more
efficient to do:

    git log HEAD~15

than

    git log --max-count=15

Profiling data is available on #686 (specifically
https://github.com/magit/magit/pull/686#issuecomment-20141371).