remove magit-get-remote, sanitize magit-{push,pull}
Until a short while ago Magit had very poor support for setting
the (remote or local) branch tracked by a branch.  Actually the
only way to do so was as a side-effect of `magit-push'.  We now
have proper support for setting the tracked branch and no longer
"need" `magit-push' and `magit-pull' to fallback to "origin" to
make dealing with this less painful (and potentially dangerous).

Nuke `magit-get-remote' which is where the now unnecessary
fallback was implemented.