magit-section-when: new macro
Add new macro `magit-section-when' which unlike `magit-section-case'
cannot be used to dispatch based on section type.  Instead it used
to match a single section type (or a single set of section types),
and therefor serves well as a predicate.

Like `when' and for that matter `if', `magit-section-when' is not
strictly necessary.  But it is semantically more clear and, especially
it is further optimized for how it is commonly used, also less verbose.