magit-discard: act on staged hunks despite unstaged changes
`git apply --index' refuses to apply any patches if the file to be
patched is not up-to-date.  But in many cases that would work without
any conflicts.  So when there are both stages as well as unstaged
changes and the user requests that a staged change is discarded,
first discard it in the index only using `--cached' and then also
attempt to do so to the working tree.  If the second step fails
due to conflicts, then the result is equivalent to unstaging,
which isn't so bad.