magit-{,un}stage-file: new commands
Remove the ability to stage something other than the change at point
from `magit-stage' and instead add the new command `magit-stage-file',
which stages the file at point, if any, and else prompts for one.

Likewise create `magit-unstage-file'.  (Since `magit-unstage' never
allowed picking a file in the minibuffer, we don't have to touch that.)