reduce nesting inside "Apply" section
Section "Apply" is itself a subsection of toplevel "Porcelain", and was
previously divided into "Apply Commands" and "Apply Core".  This soon
won't make sense anymore, since many functions that were located inside
the latter will be turned into commands.

Move the definition of `magit-apply' next to the other functions that
are used for "applying, in the strict *and* loose sense" (other
functions, in other subsections do e.g. "stage" changes, which isn't the
same as applying in the strict sense, but is implemented by using the
low-level "apply" functions never-the-less.  These generic apply
functions were previously located in the "Apply Core" subsection).
Also move the definition of `magit-discard-diff'.