magit-{apply-{diff,hunk},insert-{*}}: name argument SECTION
Rename the arguments DIFF and HUNK of `magit-apply-diff',
`magit-apply-hunk', and the related `magit-insert-{*}'s to SECTION.

It isn't obvious that these functions expect section objects (strings
would also make sense).  However once we know that some kind of section
is expected, it is obvious what kind of section that is.