add support for per-section-type keymaps
Teach `magit-insert-section' to detect whether a keymap is defined
for the section being inserted and if so propertize the section's text
accordingly using the `keymap' text property.  The keymap variable has
to be named `magit-TYPE-section-map'.

Also define several such section keymaps.  The bindings will be added
in subsequent commits, when we move away from `magit-section-action',
which provides a similar but less flexible and more complicated way to
accomplish the same.