no longer overload magit-cherry-pick
Previously `magit-cherry-pick' did cherry-pick the commit at point, or
if the section at point was a stash it did pop that stash.  Popping a
stash has nothing to do with cherry-picking, and since the name of this
function rather explicitly says that it does cherry-pick it should not
pop stashes too.  But it does indeed make sense to use the same key to
invoke these two technically unrelated actions, because from a user
perspective they do something "similar".

Replace the old `magit-cherry-pick' with what `magit-cherry-pick-commit'
used to do, but be a bit smarter about it.  Remove the latter.

Use the new "section keymap" functionality added in the previous commit
a first time.  Pressing "A" on a stash no longer uses the buffer-wide
binding `magit-cherry-pick' but instead uses `magit-stash-pop', as bound
in the new `magit-stash-section-map'.  Also bind `magit-cherry-pick' in
the new `magit-commit-section-map', but that is actually redundant.