sanitize reset commands
Add a dedicated command for each reset variant; index, head, soft, and
hard.

Remove `magit-unstage-all' (optionally asks for confirmation) in favor
of the new `magit-reset-index' (asks for a commit, defaulting to the
commit at point).

In `magit-unstage' do not replace `magit-unstage-all' with
`magit-reset-index' because there is no way to keep it from asking for
a commit, effectively asking for confirmation.  Instead move the code
previously in `magit-unstage-all' directly into `magit-unstage'.