magit-stage-modified: new command, replacing magit-stage-all
The old command was badly named.  It uses `git add', which happens to
have an argument `--all', but the command did not use that (except with
a prefix argument), it used neither `--all' nor `--update'.

The new command uses `--update', so now deleted files are also removed,
in addition to changes to modified files being added.  Because the new
command also takes a prefix argument which means to use `--all' it is
not named `magit-stage-update' but `magit-stage-modified', which is
sufficiently distinct from the names of the arguments to `git add',
which might or might not be used.