magit-wash-diffstat: use preliminary filenames as value
The filenames that Git prints when --stat is used may be abbreviated.
For that reason we set the values of the file subsections of the
`diffstat' section once we parse the corresponding file section that
is part of the actual diff.  However when showing a merge commit it
is possible for diffstat information to be displayed even though
there is no diff.  Therefor set the value of the diffstat subsection
when it is created and possibly change that value if there actually
is a corresponding diff.  Setting the proper filename can still fail
when a filename is actually is abbreviated and there *also* is no
corresponding diff.  Previously it always failed when there was no
diff.  Furthermore the error displayed when it actually does happen
is more useful now.