magit-insert-child-count: remove trailing colon
A colon at the and of a heading still indicates that the number of child
sections should be shown at the end of the heading.  And so the strings
passed to most calls to `magit-insert-heading' still end with a colon.

However `magit-insert-child-count' now replaces the colon with " (N)"
instead of inserting that that string before the colon.