restore behavior of "x" on a commit
Commit 11dc61ec479d671510e4c18ac73dfe096ad89518 intentionally changed
the behavior of pressing "x" on a commit.  Starting with that "x" did a
hard reset instead of a mixed reset, because "U" already does a mixed
reset.  Unfortunately many users would not notice this change until
they had already nuked some uncommitted changes.  So let's revert to
the old behavior.  "x" does a mixed reset again and "C-u x" has to be
used for a hard reset.

Add the new command `magit-revert' which behaves like
`magit-revert-head' used to do and bind "x" to it.  Locally I revert
this revert by binding "x" to `magit-revert-hard' again.