magit-stash-{apply,pop}: always respect magit-stash-popup-defaults
When `magit-stash-apply' and `magit-stash-pop' are called directly
(as opposed to from the popup) then explicitly check whether
`magit-stash-popup-defaults' contains "--index" and if so use it.

This is only a kludge and should be generalized, for several other
commands doing this would also make sense, but here it is especially
annoying that this feature doesn't exist yet, thus this kludge.