default to use per-repository buffers
Until now all buffers except for the status buffers where shared across
all repositories.  It was already possible to use per-repository buffers
for other modes too, but that was undocumented.

Replace all `magit-X-buffer-name' variables with properly documented
`magit-X-buffer-name-format' options and default to inserting the
absolute path to the toplevel of the current repository.

Also teach `magit-mode-get-buffer' what to do when `default-directory'
isn't below a git repository.