simplify moving point outside invisible text
This should not be necessary at all anymore, and will be removed
completely in the next commit.  There is a slight risk that I am
wrong about this so I am adding this commit so we have something
more civilized to go back to than the old implementation in case
we need an emergency fix.

See next commit for why we needed this and why we don't do so
anymore.