let the command loop move point outside of invisible text
The reason we needed this kludge was that process sentinels which may
cause point to move inside invisible text and which also have to be
considered part of some command did not actually run until after the
command returned.  The editor command loop then went on to do its point
adjustment and then run "threads", including the sentinel of the process
started by the command which just returned.

These sentinels were likely to cause point to end up inside invisible
text, but when they ran it was to late for the command loops point
adjustment to kick in.

We used to work around that by using `post-command-hook' which happens
to be run even later than pending "threads".  Running such sentinels
after the command has already returned caused other issues besides this
one and so that was fixed by waiting inside the command for the process
to return.

This is the right thing to do because the only reason why asynchronous
processes are used in the first place is that this allows sending input
to the process.  But as far as Magit is concerned these processes should
be synchronous.  Now we wait, making at least the commands "synchronous"
and this kludge unnecessary.